0.6.0:     - CHANGED  : config file format changed     - CHANGED  : now installed packages are stored in 'installed' meta repository     - CHANGED  : alias files have extension changed form *.njc to *.cmd-alias.json     - CHANGED  : now nuts looks for system env variable NUTS_WORKSPACE for default workspace location     - ADDED    : support for Interrupting Copy     - ADDED    : support to ps (list processes)     - ADDED    : support progress options     - CHANGED  : worky, searches now for modified deployments with same version but different content     - FIXED    : encoding problem with json/xml

+++ nuts-core/src/main/java/net/vpc/app/nuts/runtime/io/DefaultNutsIOCopyAction.java
@@ -561,7 +561,7 @@ public class DefaultNutsIOCopyAction implements NutsIOCopyAction {
                         }
                     }
                 } finally {
-                    if (temp != null) {
+                    if (temp != null && Files.exists(temp)) {
                         Files.delete(temp);
                     }
                 }
